Now Protect Mazaalai
The Gobi bear, Ursus arctos gobiensis (known in Mongolian as the mazaalai/Мазаалай), is a subspecies of the brown bear, Ursus arctos, that is found in the Gobi Desert of Mongolia. It is listed as critically endangered by Mongolian Redbook of Endangered Species and by the Zoological Society of London using IUCN standards. The UN backed Convention of Migratory Species selected Gobi bear for protection in 2017.[1] The population included only around 30 adults in 2009[2] and is separated by enough distance from other brown bear populations to achieve reproductive isolation.