National Peoples Movement
The Social, Economic and Political Crisis in Sri Lanka is rapidly escalating and the structure of the network of government institutes is fast disintegrating leading to chaos. Meanwhile the inconsistencies and the paradoxes between the Executive and the Legislature are further aggravating the crisis while the country is at a standstill. The acceptance and respect for politicians and political parties among the general public is rapidly deteriorating.
Present power struggle and rivalry action among main Political Parties may eventually pave the way for the evolution of a brutal dictatorship, in the absence of a feasible alternative to the people of this country. More and more it is becoming clear that this country is at the bottom of the deterioration and can be saved only through a structured and an organized Peoples’ movement, which understands the gravity of the situation.
Therefore in end of 2018, we firmly expressed our active engagement in building the “National Peoples  Movement”  (NPM) with conscious awareness, determination and commitment to synergize and network Progressive Organizations and Movements, Trade Unions, Farmer and Fisher communities,  Intellects, Artists, Literati, Entrepreneurs, Environmental, Civil, Political and Social Organizations and Activists etc., that are representing all Ethnic communities, Cultural and Religious Groups, Youth, Women, and people of all walks of life, who are aspiring to build an united and inclusive “Sri Lankan Nation”  to overcome the present crisis of governance and the most likely state of anarchy to follow and create a better country for the future generations.
As equal members “National Peoples Movement” comprises 20 island wide civil society groups which represent different roles in their social and political activism. Deshodaya Movement and United Professionals’ Movement are leading the way of NPM.